requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
docker-pycreds==0.4.0
faiss-gpu==1.7.2
filelock==3.13.1
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
googledrivedownloader==1.1.0
huggingface-hub==0.30.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
littleutils==0.2.4
lmdb==1.6.2
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.3
numpy==1.26.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
ogb==1.3.6
outdated==0.2.2
packaging==25.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyg-lib==0.4.0+pt21cu118
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentry-sdk==2.27.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.1.0+cu118
torch-cluster==1.6.3+pt21cu118
torch-scatter==2.1.2+pt21cu118
torch-sparse==0.6.18+pt21cu118
torch-spline-conv==1.2.2+pt21cu118
torch_geometric==2.4.0
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.67.1
transformers==4.51.3
triton==2.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
yacs==0.1.8
yarl==1.20.0
